[issue-22] Update resume and linked it to google drive
Removed resume component from app because it was not maintainable. The css necessary to create a web resume was messy and it also required to keep it updated in 2 locations, word document and web. To improve making only one change, I decided to link the resume from my google drive as it is where I store my updated resume.

issue: https://github.com/drusc0/web-app/issues/22